requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
entrypoints==0.4
faiss-cpu==1.7.4
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.4.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.38.0
huggingface-hub==0.11.1
idna==3.10
importlib_metadata==8.5.0
InstructorEmbedding==1.0.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.0.284
langsmith==0.0.92
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==1.23.5
openai==0.27.0
packaging==24.2
pandas==1.5.3
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==3.20.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.28.2
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.1.3
scipy==1.14.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
streamlit==1.22.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==1.0.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.25.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==1.26.20
validators==0.34.0
watchdog==6.0.0
yarl==1.18.3
zipp==3.21.0
